moved code to client.py, created models.py
