+++ ingestion/src/metadata/ingestion/source/dashboard/metabase/client.py
@@ -11,6 +11,8 @@
 """
 REST Auth & Client for Metabase
 """
+from typing import List, Optional
+import traceback
 import requests
 from metadata.ingestion.source.dashboard.dashboard_service import DashboardServiceSource
 from metadata.generated.schema.metadataIngestion.workflow import (
@@ -19,6 +21,9 @@ from metadata.generated.schema.metadataIngestion.workflow import (
 from metadata.generated.schema.entity.services.connections.metadata.openMetadataConnection import (
     OpenMetadataConnection,
 )
+from metadata.utils.logger import ingestion_logger
+
+logger = ingestion_logger()
 
 class MetabaseClient(DashboardServiceSource):
     
@@ -26,10 +31,14 @@ class MetabaseClient(DashboardServiceSource):
         config: WorkflowSource,
         metadata_config: OpenMetadataConnection
     ):
-        super().__init__(config, metadata_config)
-        self.metabase_session = self.client["metabase_session"]
+        try:
+          super().__init__(config, metadata_config)
+          self.metabase_session = self.client["metabase_session"]
+        except Exception as exc:
+            logger.debug(traceback.format_exc())
+            logger.error(f"Error in initializing Metabase Client: {exc}")
     
-    def req_get(self, path):
+    def req_get(self, path: str) -> requests.Response:
       """Send get request method
 
       Args:
@@ -40,3 +49,39 @@ class MetabaseClient(DashboardServiceSource):
           headers=self.metabase_session,
           timeout=30,
       )
+    
+    def get_dashboards_list(self) -> Optional[List[dict]]:
+        """
+        Get List of all dashboards
+        """
+        resp_dashboards = self.req_get("/api/dashboard")
+        if resp_dashboards.status_code == 200:
+            return resp_dashboards.json()
+        return []
+    
+    def get_dashboard_name(self, dashboard: dict) -> str:
+        """
+        Get Dashboard Name
+        """
+        return dashboard["name"]
+
+    def get_dashboard_details(self, dashboard: dict) -> dict:
+        """
+        Get Dashboard Details
+        """
+        resp_dashboard = self.req_get(f"/api/dashboard/{dashboard['id']}")
+        return resp_dashboard.json()
+    
+    def get_database(self, database_id: str) -> dict:
+        """
+        Get Database using database ID
+        """
+        resp_database = self.req_get(f"/api/database/{database_id}")
+        return resp_database.json()
+    
+    def get_table(self, table_id: str) -> dict:
+        """
+        Get Table using table ID
+        """
+        resp_table = self.req_get(f"/api/table/{table_id}")
+        return resp_table.json()
