+++ ingestion/src/metadata/ingestion/source/dashboard/metabase/metadata.py
@@ -51,16 +51,11 @@ class MetabaseSource(DashboardServiceSource):
     Metabase Source Class
     """
 
-    config: WorkflowSource
-    metadata_config: OpenMetadataConnection
-
     def __init__(
         self,
         config: WorkflowSource,
         metadata_config: OpenMetadataConnection,
     ):
-        super().__init__(config, metadata_config)
-        self.metabase_session = self.client["metabase_session"]
         self.client = MetabaseClient(config, metadata_config)
 
     @classmethod
@@ -72,28 +67,6 @@ class MetabaseSource(DashboardServiceSource):
                 f"Expected MetabaseConnection, but got {connection}"
             )
         return cls(config, metadata_config)
-
-    def get_dashboards_list(self) -> Optional[List[dict]]:
-        """
-        Get List of all dashboards
-        """
-        resp_dashboards = self.client.req_get("/api/dashboard")
-        if resp_dashboards.status_code == 200:
-            return resp_dashboards.json()
-        return []
-
-    def get_dashboard_name(self, dashboard: dict) -> str:
-        """
-        Get Dashboard Name
-        """
-        return dashboard["name"]
-
-    def get_dashboard_details(self, dashboard: dict) -> dict:
-        """
-        Get Dashboard Details
-        """
-        resp_dashboard = self.client.req_get(f"/api/dashboard/{dashboard['id']}")
-        return resp_dashboard.json()
 
     def yield_dashboard(
         self, dashboard_details: dict
@@ -219,78 +192,36 @@ class MetabaseSource(DashboardServiceSource):
     def _yield_lineage_from_query(
         self, chart_details: dict, db_service_name: str, dashboard_name: str
     ) -> Optional[AddLineageRequest]:
-        resp_database = self.client.req_get(f"/api/database/{chart_details['database_id']}")
-        if resp_database.status_code == 200:
-            database = resp_database.json()
-            query = (
-                chart_details.get("dataset_query", {})
-                .get("native", {})
-                .get("query", "")
+        database = self.client.get_database(chart_details['database_id'])
+        query = (
+            chart_details.get("dataset_query", {})
+            .get("native", {})
+            .get("query", "")
+        )
+        lineage_parser = LineageParser(query)
+        for table in lineage_parser.source_tables:
+            database_schema_name, table = fqn.split(str(table))[-2:]
+            database_schema_name = (
+                None
+                if database_schema_name == "<default>"
+                else database_schema_name
             )
-            lineage_parser = LineageParser(query)
-            for table in lineage_parser.source_tables:
-                database_schema_name, table = fqn.split(str(table))[-2:]
-                database_schema_name = (
-                    None
-                    if database_schema_name == "<default>"
-                    else database_schema_name
-                )
-                database = database.get("details", {}).get("db", None)
-                if database:
-                    from_entities = search_table_entities(
-                        metadata=self.metadata,
-                        database=database,
-                        service_name=db_service_name,
-                        database_schema=database_schema_name,
-                        table=table,
-                    )
-                else:
-                    from_entities = search_table_entities(
-                        metadata=self.metadata,
-                        service_name=db_service_name,
-                        database=None,
-                        database_schema=database_schema_name,
-                        table=table,
-                    )
-
-                to_fqn = fqn.build(
-                    self.metadata,
-                    entity_type=LineageDashboard,
-                    service_name=self.config.serviceName,
-                    dashboard_name=dashboard_name,
-                )
-                to_entity = self.metadata.get_by_name(
-                    entity=LineageDashboard,
-                    fqn=to_fqn,
-                )
-
-                for from_entity in from_entities:
-                    yield self._get_add_lineage_request(
-                        to_entity=to_entity, from_entity=from_entity
-                    )
-
-    def _yield_lineage_from_api(
-        self, chart_details: dict, db_service_name: str, dashboard_name: str
-    ) -> Optional[AddLineageRequest]:
-        resp_tables = self.client.req_get(f"/api/table/{chart_details['table_id']}")
-        if resp_tables.status_code == 200:
-            table = resp_tables.json()
-            database_name = table.get("db", {}).get("details", {}).get("db", None)
-            if database_name:
+            database = database.get("details", {}).get("db", None)
+            if database:
                 from_entities = search_table_entities(
                     metadata=self.metadata,
-                    database=database_name,
+                    database=database,
                     service_name=db_service_name,
-                    database_schema=table.get("schema"),
-                    table=table.get("display_name"),
+                    database_schema=database_schema_name,
+                    table=table,
                 )
             else:
                 from_entities = search_table_entities(
                     metadata=self.metadata,
                     service_name=db_service_name,
                     database=None,
-                    database_schema=table.get("schema"),
-                    table=table.get("display_name"),
+                    database_schema=database_schema_name,
+                    table=table,
                 )
 
             to_fqn = fqn.build(
@@ -308,3 +239,41 @@ class MetabaseSource(DashboardServiceSource):
                 yield self._get_add_lineage_request(
                     to_entity=to_entity, from_entity=from_entity
                 )
+
+    def _yield_lineage_from_api(
+        self, chart_details: dict, db_service_name: str, dashboard_name: str
+    ) -> Optional[AddLineageRequest]:
+        table = self.client.get_table(chart_details['table_id'])
+        database_name = table.get("db", {}).get("details", {}).get("db", None)
+        if database_name:
+            from_entities = search_table_entities(
+                metadata=self.metadata,
+                database=database_name,
+                service_name=db_service_name,
+                database_schema=table.get("schema"),
+                table=table.get("display_name"),
+            )
+        else:
+            from_entities = search_table_entities(
+                metadata=self.metadata,
+                service_name=db_service_name,
+                database=None,
+                database_schema=table.get("schema"),
+                table=table.get("display_name"),
+            )
+
+        to_fqn = fqn.build(
+            self.metadata,
+            entity_type=LineageDashboard,
+            service_name=self.config.serviceName,
+            dashboard_name=dashboard_name,
+        )
+        to_entity = self.metadata.get_by_name(
+            entity=LineageDashboard,
+            fqn=to_fqn,
+        )
+
+        for from_entity in from_entities:
+            yield self._get_add_lineage_request(
+                to_entity=to_entity, from_entity=from_entity
+            )
